add exam file

+++ abstract-machine/am/src/platform/nemu/ioe/gpu.c
@@ -1,6 +1,6 @@
 #include <am.h>
-#include <stdio.h>
 #include <nemu.h>
+
 #define SYNC_ADDR (VGACTL_ADDR + 4)
 void __am_gpu_config(AM_GPU_CONFIG_T *cfg);
 AM_GPU_CONFIG_T cfg;
@@ -19,12 +19,11 @@ void __am_gpu_config(AM_GPU_CONFIG_T *cfg) {
 
 void __am_gpu_fbdraw(AM_GPU_FBDRAW_T *ctl) {
   // 开始输出画面
-  uint32_t *fb = (uint32_t *)FB_ADDR;
+  uint32_t *fb = (uint32_t *)(uintptr_t)FB_ADDR;
   for (int i = 0;i < ctl->h;++ i) {
     int y = i + ctl->y;
     for (int j = 0;j < ctl->w;++ j) {
       int x = j + ctl->x;
-      //printf("%d %d\n", x, y);
       fb [y * cfg.width + x] = ((uint32_t*)(ctl->pixels))[i * ctl->w + j];
     }
   }
